requirements.txt
# Flask Core
Flask==3.1.2 #
Flask-CORS==6.0.1 #
Flask-Limiter==4.0.0 #

# Database
Flask-SQLAlchemy==3.1.1 #
SQLAlchemy==2.0.44 #
pyodbc==5.3.0 #

# Authentication & Security
Flask-JWT-Extended==4.7.1 #
PyJWT==2.10.1 #
bcrypt==5.0.0 #

# API Documentation
flasgger==0.9.7.1 #

# Testing
pytest==9.0.1 #
pytest-cov==7.0.0 #
pytest-flask==1.3.0 #
pytest-mock==3.15.1 #
coverage==7.12.0 #

# Development
Faker==22.0.0

# Utilities
python-dateutil==2.9.0 #

# Rate Limiting Backend
limits==3.7.0

# WSGI Server for Production
gunicorn==21.2.0
waitress==3.0.0

# Logging
python-json-logger==2.0.7

# Environment
environs==14.5.0 #